refactor([issue-745]): split taskLearning.js into store/metrics/routing submodules
Separate the 1939-LOC monolith into a server/services/taskLearning/
directory split by concern: a shared persistence store, metrics
aggregation (recording + aggregate rebuilds), heuristic routing
(cooldown/skip/confidence), duration estimates, insights/recommendations,
prompt recommendations, and lifecycle wiring. taskLearning.js becomes a
barrel re-exporting the identical public API, so every import site is
unchanged. Pure structural move — behavior preserved, all 31 exports
verified identical.
